Move `enums` module into `model` module.
