+++ src/model/context.rs
@@ -5,7 +5,7 @@ use std::collections::HashMap;
 use super::device::Device;
 use super::track::FullTrack;
 use super::PlayingItem;
-use crate::enums::{CurrentlyPlayingType, DisallowKey, RepeatState, Type};
+use crate::model::{CurrentlyPlayingType, DisallowKey, RepeatState, Type};
 /// Context object
 /// [get the users currently playing track](https://developer.spotify.com/web-api/get-the-users-currently-playing-track/)
 #[derive(Clone, Debug, Serialize, Deserialize)]
