+++ src/model/mod.rs
@@ -6,6 +6,7 @@ pub mod category;
 pub mod context;
 pub mod cud_result;
 pub mod device;
+pub mod enums;
 pub mod image;
 pub mod offset;
 pub mod page;
@@ -27,7 +28,7 @@ pub enum PlayingItem {
 }
 
 pub use {
-    album::*, artist::*, audio::*, category::*, context::*, cud_result::*, device::*, image::*,
-    offset::*, page::*, playing::*, playlist::*, recommend::*, search::*, show::*, track::*,
-    user::*,
+    album::*, artist::*, audio::*, category::*, context::*, cud_result::*, device::*, enums::*,
+    image::*, offset::*, page::*, playing::*, playlist::*, recommend::*, search::*, show::*,
+    track::*, user::*,
 };
